Add submitDisabled to reducer state.
